green_agent/pyproject.toml
[project]
name = "healthcare-fraud-green-agent"
version = "0.1.0"
description = "AgentBeats Green Agent for Healthcare Claims Fraud Detection"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    # Environment dependencies
    "fastapi>=0.111.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
